unified modify interations (fan, corridor, orbit).
